requirements.txt
Flask
gunicorn
google-cloud-storage
google-cloud-datastore